Crate srml_indices

An index is a short form of an address. This module handles allocation of indices for a newly created accounts.

Modules

address
Address type that is union of index and id for an account.

Structs

EnumSet
The enumeration sets.

GenesisConfig

Module

NextEnumSet
The next free enumeration set.

SimpleResolveHint
Simple encode-based resolve hint implementation.

Enums

Call

RawEvent
Events for this module.

Traits

ResolveHint
Turn an Id into an Index, or None for the purpose of getting a hint at a possibly desired index.

Trait
The module's config trait.

Type Definitions

Address

Event
RawEvent specialized for the configuration Trait